Initial commit
Prepared the library with an initial 1.0 version. Added a readme to start talking about the library.

Still need to figure out why coverage is not being calculated. Also need to setup the build pipeline and CI pipeline to release the package on github packages.

Will need contributing guidelines if we want to allow PRs.