Finish the improved rendered and fixed tests
Testify now owns its own logger and will render the App component. The tests do not use the testify function and instead call the App manually. All the core logic for testify has been moved to the app.

Testify will build its own logger when started so calling it multiple times will not cause the logger to break to contain invalid data. This also allows us to not render anything during the test so they are fully self-contained and will not modify the test framework (Since testify tests itself).