Modernize the library so it works with modern node without hacks
The library was built a few years ago when modules weren't as easy to work with in node as they are today, especially in node 14. This PR updates the build and testing logic so we can finally use the modern versions of node without any ESM shenanigans. It should be a lot easier to use and test, and somewhat faster to run on modern node.

Also updated codecov and its badge so we get the modern coverage with the correct uploader. V1 has been disabled for more than a year now.